requirements.txt
alembic==1.13.1 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.9" and python_version < "4.0"
asgiref==3.8.1 ; python_version >= "3.9" and python_version < "4.0"
cachetools==5.3.3 ; python_version >= "3.9" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
exceptiongroup==1.2.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.70.1 ; python_version >= "3.9" and python_version < "4.0"
google-api-core==2.18.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-core[grpc]==2.18.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.29.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-bigquery==3.19.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.7.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.63.0 ; python_version >= "3.9" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.9" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
grpcio-status==1.62.1 ; python_version < "4.0" and python_version >= "3.9"
grpcio==1.62.1 ; python_version < "4.0" and python_version >= "3.9"
gunicorn==20.1.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.6 ; python_version >= "3.9" and python_version < "4.0"
loguru==0.5.3 ; python_version >= "3.9" and python_version < "4.0"
mako==1.3.2 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.9" and python_version <= "3.11" or python_version >= "3.12" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.25.3 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.4.0 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.14 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.9" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
setuptools==69.2.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
sqlalchemy==1.4.52 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.16.0 ; python_version >= "3.9" and python_version < "4.0"
stripe==2.76.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.10.0 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.15.0 ; python_version >= "3.9" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
